requirements.txt
Babel==2.6.0
bcrypt==3.1.4
blinker==1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
dominate==2.3.1
Flask==1.0.2
Flask-Admin==1.5.1
Flask-BabelEx==0.9.3
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Moment==0.6.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SocketIO==3.0.1
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
passlib==1.7.1
Pillow==5.2.0
psycopg2==2.7.5
pycparser==2.18
PyJWT==1.6.4
PySocks==1.6.8
python-engineio==2.2.0
python-socketio==2.0.0
pytz==2018.5
requests==2.19.1
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.9
twilio==6.14.10
urllib3==1.23
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
